+++ src/DeMarketService/Common/HttpApiModel/ResponseModel/OrderResponse.cs
@@ -29,7 +29,7 @@ namespace deMarketService.Common.Model.HttpApiModel.ResponseModel
         /// <summary>
         /// 商品价格
         /// </summary>
-        public decimal price { get; set; }
+        public double price { get; set; }
         /// <summary>
         /// 图片
         /// </summary>
@@ -122,7 +122,7 @@ namespace deMarketService.Common.Model.HttpApiModel.ResponseModel
         /// </summary>
         public decimal price_actual
         {
-            get { return price / decimals_long; }
+            get { return (decimal)(price / (double)decimals_long); }
         }
         /// <summary>
         /// 订单总价
